refactor(repository, participant): participant repository 통합
