+++ src/main/java/com/pnu/momeet/domain/participant/service/ParticipantEntityService.java
@@ -1,7 +1,7 @@
 package com.pnu.momeet.domain.participant.service;
 
 import com.pnu.momeet.domain.participant.entity.Participant;
-import com.pnu.momeet.domain.participant.repository.ParticipantDslRepository;
+import com.pnu.momeet.domain.participant.repository.ParticipantDslRepositoryImpl;
 import com.pnu.momeet.domain.participant.repository.ParticipantRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -19,7 +19,6 @@ import java.util.function.Consumer;
 @RequiredArgsConstructor
 public class ParticipantEntityService {
     private final ParticipantRepository participantRepository;
-    private final ParticipantDslRepository participantDslRepository;
 
     @Transactional(readOnly = true)
     public List<Participant> getAllByMeetupId(UUID meetupId) {
@@ -28,7 +27,7 @@ public class ParticipantEntityService {
 
     @Transactional(readOnly = true)
     public List<Participant> findAllVisibleByMeetupId(UUID meetupId, UUID viewerMemberId) {
-        return participantDslRepository.findAllVisibleByMeetupId(meetupId, viewerMemberId);
+        return participantRepository.findAllVisibleByMeetupId(meetupId, viewerMemberId);
     }
 
     @Transactional(readOnly = true)
